added sleep anywhere feature, fixed onLegUse being called in cases where it doesnt make sense, nerfed magic healing, fixed some translations, added ui glow color config option

+++ src/main/java/net/zaharenko424/casualties_cubed/client/ticksounds/PainDrone.java
@@ -45,7 +45,7 @@ public class PainDrone extends AbstractTickableSoundInstance {
 
     public void update(Player player){
         player.getCapability(PlayerHealthProvider.PLAYER_HEALTH_DATA).ifPresent(h->{
-            float painscale = (float) ((h.getAveragePain())/100f);
+            float painscale = (h.getAveragePain())/100f;
             float cons = h.getConsciousness();
             painscale = Mth.clamp(painscale,0.6f,1.1f);
             if (h.getAveragePain() < 50) {
